# Iterators: re-implement mergeEntryIterator using loser.Tree for performance