--- a/pkg/iter/entry_iterator.go
+++ b/pkg/iter/entry_iterator.go
@@ -1,7 +1,6 @@
 package iter
 
 import (
-	"container/heap"
 	"context"
 	"io"
 	"math"
@@ -57,39 +56,6 @@ func (i *streamIterator) Close() error {
 	return nil
 }
 
-type iteratorHeap []EntryIterator
-
-func (h iteratorHeap) Len() int            { return len(h) }
-func (h iteratorHeap) Swap(i, j int)       { h[i], h[j] = h[j], h[i] }
-func (h iteratorHeap) Peek() EntryIterator { return h[0] }
-func (h *iteratorHeap) Push(x interface{}) {
-	*h = append(*h, x.(EntryIterator))
-}
-
-func (h *iteratorHeap) Pop() interface{} {
-	old := *h
-	n := len(old)
-	x := old[n-1]
-	*h = old[0 : n-1]
-	return x
-}
-
-type iteratorSortHeap struct {
-	iteratorHeap
-	byAscendingTime bool
-}
-
-func (h iteratorSortHeap) Less(i, j int) bool {
-	t1, t2 := h.iteratorHeap[i].Entry().Timestamp.UnixNano(), h.iteratorHeap[j].Entry().Timestamp.UnixNano()
-	if t1 == t2 {
-		return h.iteratorHeap[i].StreamHash() < h.iteratorHeap[j].StreamHash()
-	}
-	if h.byAscendingTime {
-		return t1 < t2
-	}
-	return t1 > t2
-}
-
 // HeapIterator iterates over a heap of iterators with ability to push new iterators and get some properties like time of entry at peek and len
 // Not safe for concurrent use
 type HeapIterator interface {
@@ -101,16 +67,8 @@ type HeapIterator interface {
 
 // mergeEntryIterator iterates over a heap of iterators and merge duplicate entries.
 type mergeEntryIterator struct {
-	heap interface {
-		heap.Interface
-		Peek() EntryIterator
-	}
-	is []EntryIterator
-	// pushBuffer contains the list of iterators that needs to be pushed to the heap
-	// This is to avoid allocations.
-	pushBuffer []EntryIterator
-	prefetched bool
-	stats      *stats.Context
+	tree  *loser.Tree[sortFields, EntryIterator]
+	stats *stats.Context
 
 	// buffer of entries to be returned by Next()
 	// We buffer entries with the same timestamp to correctly dedupe them.
@@ -124,104 +82,64 @@ type mergeEntryIterator struct {
 // This means using this iterator with a single iterator will result in the same result as the input iterator.
 // If you don't need to deduplicate entries, use `NewSortEntryIterator` instead.
 func NewMergeEntryIterator(ctx context.Context, is []EntryIterator, direction logproto.Direction) HeapIterator {
-	result := &mergeEntryIterator{is: is, stats: stats.FromContext(ctx)}
-	switch direction {
-	case logproto.BACKWARD:
-		result.heap = &iteratorSortHeap{iteratorHeap: make([]EntryIterator, 0, len(is)), byAscendingTime: false}
-	case logproto.FORWARD:
-		result.heap = &iteratorSortHeap{iteratorHeap: make([]EntryIterator, 0, len(is)), byAscendingTime: true}
-	default:
-		panic("bad direction")
-	}
-
+	maxVal, less := treeLess(direction)
+	result := &mergeEntryIterator{stats: stats.FromContext(ctx)}
+	result.tree = loser.New(is, maxVal, sortFieldsAt, less, result.closeEntry)
 	result.buffer = make([]entryWithLabels, 0, len(is))
-	result.pushBuffer = make([]EntryIterator, 0, len(is))
 	return result
 }
 
-// prefetch iterates over all inner iterators to merge together, calls Next() on
-// each of them to prefetch the first entry and pushes of them - who are not
-// empty - to the heap
-func (i *mergeEntryIterator) prefetch() {
-	if i.prefetched {
-		return
-	}
-
-	i.prefetched = true
-	for _, it := range i.is {
-		i.requeue(it, false)
-	}
-
-	// We can now clear the list of input iterators to merge, given they have all
-	// been processed and the non empty ones have been pushed to the heap
-	i.is = nil
-}
-
-// requeue pushes the input ei EntryIterator to the heap, advancing it via an ei.Next()
-// call unless the advanced input parameter is true. In this latter case it expects that
-// the iterator has already been advanced before calling requeue().
-//
-// If the iterator has no more entries or an error occur while advancing it, the iterator
-// is not pushed to the heap and any possible error captured, so that can be get via Error().
-func (i *mergeEntryIterator) requeue(ei EntryIterator, advanced bool) {
-	if advanced || ei.Next() {
-		heap.Push(i.heap, ei)
-		return
-	}
-
-	if err := ei.Error(); err != nil {
+func (i *mergeEntryIterator) closeEntry(e EntryIterator) {
+	if err := e.Error(); err != nil {
 		i.errs = append(i.errs, err)
 	}
-	util.LogError("closing iterator", ei.Close)
+	util.LogError("closing iterator", e.Close)
+
 }
 
 func (i *mergeEntryIterator) Push(ei EntryIterator) {
-	i.requeue(ei, false)
+	i.tree.Push(ei)
 }
 
-// Next pop iterators from the heap until it finds an entry with a different timestamp or stream hash.
-// For each iterators we also buffer entries with the current timestamp and stream hash deduping as we loop.
-// If the iterator is not fully exhausted, it is pushed back to the heap.
+// Next fetches entries from the tree until it finds an entry with a different timestamp or stream hash.
+// Generally i.buffer has one or more entries with the same timestamp and stream hash,
+// followed by one more item where the timestamp or stream hash was different.
 func (i *mergeEntryIterator) Next() bool {
-	i.prefetch()
-
-	if len(i.buffer) != 0 {
-		i.nextFromBuffer()
-		return true
+	if len(i.buffer) < 2 {
+		i.fillBuffer()
 	}
-
-	if i.heap.Len() == 0 {
+	if len(i.buffer) == 0 {
 		return false
 	}
+	i.nextFromBuffer()
 
-	// shortcut for the last iterator.
-	if i.heap.Len() == 1 {
-		i.currEntry.Entry = i.heap.Peek().Entry()
-		i.currEntry.labels = i.heap.Peek().Labels()
-		i.currEntry.streamHash = i.heap.Peek().StreamHash()
+	return true
+}
 
-		if !i.heap.Peek().Next() {
-			i.heap.Pop()
-		}
-		return true
+func (i *mergeEntryIterator) fillBuffer() {
+	if !i.tree.Next() {
+		return
 	}
+	// At this point we have zero or one items in i.buffer, and the next item is available from i.tree.
 
 	// We support multiple entries with the same timestamp, and we want to
-	// preserve their original order. We look at all the top entries in the
-	// heap with the same timestamp, and pop the ones whose common value
-	// occurs most often.
-Outer:
-	for i.heap.Len() > 0 {
-		next := i.heap.Peek()
+	// preserve their original order.
+	// Entries with identical timestamp and line are removed as duplicates.
+	for {
+		next := i.tree.Winner()
 		entry := next.Entry()
-		if len(i.buffer) > 0 &&
+		previous := i.buffer
+		i.buffer = append(i.buffer, entryWithLabels{
+			Entry:      entry,
+			labels:     next.Labels(),
+			streamHash: next.StreamHash(),
+		})
+		if len(i.buffer) > 1 &&
 			(i.buffer[0].streamHash != next.StreamHash() ||
 				!i.buffer[0].Entry.Timestamp.Equal(entry.Timestamp)) {
 			break
 		}
 
-		heap.Pop(i.heap)
-		previous := i.buffer
 		var dupe bool
 		for _, t := range previous {
 			if t.Entry.Line == entry.Line {
@@ -230,52 +148,24 @@ Outer:
 				break
 			}
 		}
-		if !dupe {
-			i.buffer = append(i.buffer, entryWithLabels{
-				Entry:      entry,
-				labels:     next.Labels(),
-				streamHash: next.StreamHash(),
-			})
+		if dupe {
+			i.buffer = previous
 		}
-	inner:
-		for {
-			if !next.Next() {
-				continue Outer
-			}
-			entry := next.Entry()
-			if next.StreamHash() != i.buffer[0].streamHash ||
-				!entry.Timestamp.Equal(i.buffer[0].Entry.Timestamp) {
-				break
-			}
-			for _, t := range previous {
-				if t.Entry.Line == entry.Line {
-					i.stats.AddDuplicates(1)
-					continue inner
-				}
-			}
-			i.buffer = append(i.buffer, entryWithLabels{
-				Entry:      entry,
-				labels:     next.Labels(),
-				streamHash: next.StreamHash(),
-			})
+		if !i.tree.Next() {
+			break
 		}
-		i.pushBuffer = append(i.pushBuffer, next)
 	}
-
-	for _, ei := range i.pushBuffer {
-		heap.Push(i.heap, ei)
-	}
-	i.pushBuffer = i.pushBuffer[:0]
-
-	i.nextFromBuffer()
-
-	return true
 }
 
 func (i *mergeEntryIterator) nextFromBuffer() {
 	i.currEntry.Entry = i.buffer[0].Entry
 	i.currEntry.labels = i.buffer[0].labels
 	i.currEntry.streamHash = i.buffer[0].streamHash
+	if len(i.buffer) == 2 {
+		i.buffer[0] = i.buffer[1]
+		i.buffer = i.buffer[:1]
+		return
+	}
 	if len(i.buffer) == 1 {
 		i.buffer = i.buffer[:0]
 		return
@@ -305,26 +195,27 @@ func (i *mergeEntryIterator) Error() error {
 }
 
 func (i *mergeEntryIterator) Close() error {
-	for i.heap.Len() > 0 {
-		if err := i.heap.Pop().(EntryIterator).Close(); err != nil {
-			return err
-		}
-	}
+	i.tree.Close()
 	i.buffer = nil
-	return nil
+	return i.Error()
 }
 
 func (i *mergeEntryIterator) Peek() time.Time {
-	i.prefetch()
-
-	return i.heap.Peek().Entry().Timestamp
+	if len(i.buffer) == 0 {
+		i.fillBuffer()
+	}
+	if len(i.buffer) == 0 {
+		return time.Time{}
+	}
+	return i.buffer[0].Timestamp
 }
 
 // IsEmpty returns true if there are no more entries to pull.
 func (i *mergeEntryIterator) IsEmpty() bool {
-	i.prefetch()
-
-	return i.heap.Len() == 0
+	if len(i.buffer) == 0 {
+		i.fillBuffer()
+	}
+	return len(i.buffer) == 0
 }
 
 type entrySortIterator struct {